refactor(frontend): extract ChargesPage to src/pages/ (Phase 3)
Part of the Phase 3 migration. Route shrinks from 275 to 15 lines.

Converts the two useMemo + .filter() over useCharges().data.charges to
useCharges(params, select) calls (§3 — derived state goes through
select). Converts the useMemo deriving connectionPartnerMap from
useAccounts() to useAccounts(select). Each select result is memoized
per subscriber; the shared queryKey dedupes to a single fetch.

Extracts the inline useQuery for the balance into a useBalance hook
alongside useOpeningBalance (§3 — every query is a custom hook
returning { query, invalidate }).

Confirmation modal (reject/cancel) is kept as-is with a TODO comment;
Phase 5 will convert it to renderDrawer with a discriminated result.
